langchain_rag_engine/requirements.txt
fastapi
uvicorn
pydantic
chromadb
sentence-transformers
groq
python-dotenv
beautifulsoup4
requests
sqlalchemy
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
kagglehub
PyPDF2
PyMuPDF
xformers
python-jose[cryptography]
python-multipart
langchain
langchain-community
langchain-openai
langchain-huggingface
psutil
uvloop
cachetools
